🍫 (#22) - add navs props
