+++ nutella/src/components/Mypage/SideBar/index.tsx
@@ -1,40 +1,13 @@
 import Navigation from "./Navigation";
 import * as S from "./styled";
-import Personal from "../../../assets/icons/personal.svg";
-import Document from "../../../assets/icons/document.svg";
-import DocumentEdit from "../../../assets/icons/document_edit.svg";
-import Setting from "../../../assets/icons/setting.svg";
+import NavigationType from "../../../interface/Navigation";
+import { FC } from "react";
 
-interface NavigationType {
-  to: string;
-  text: string;
-  icon: string;
+interface PropsType {
+  navs: NavigationType[];
 }
 
-const navs: NavigationType[] = [
-  {
-    text: "프로필",
-    to: "",
-    icon: Personal,
-  },
-  {
-    text: "프로젝트",
-    to: "project",
-    icon: Document,
-  },
-  {
-    text: "보고서",
-    to: "report",
-    icon: DocumentEdit,
-  },
-  {
-    text: "계정 설정",
-    to: "setting",
-    icon: Setting,
-  },
-];
-
-const SideBar = () => {
+const SideBar: FC<PropsType> = ({ navs }) => {
   return (
     <S.Container>
       <S.NavContainer>
